# SSE: Threshold expression to use simple functions