<!--

Thank you for sending a pull request! Here are some tips:

1. If this is your first time, please read our contribution guide at https://github.com/grafana/grafana/blob/main/CONTRIBUTING.md

2. Ensure you include and run the appropriate tests as part of your Pull Request.

3. In a new feature or configuration option, an update to the documentation is necessary. Everything related to the documentation is under the docs folder in the root of the repository.

4. If the Pull Request is a work in progress, make use of GitHub's "Draft PR" feature and mark it as such.

5. If you can not merge your Pull Request due to a merge conflict, Rebase it. This gets it in sync with the main branch.

6. Name your PR as "<FeatureArea>: Describe your change", e.g. Alerting: Prevent race condition. If it's a fix or feature relevant for the changelog describe the user impact in the title. The PR title is used to auto-generate the changelog for issues marked with the "add to changelog" label.

7. If your PR content should be added to the What's New document for the next major or minor release, add the **add to what's new** label to your PR. Note that you should add this label to the main PR that introduces the feature; do not add this label to smaller PRs for the feature.

-->

**What is this feature?**
This PR updates Threshold expression to not use general-purpose math expression for evaluating the thresholds. 
Instead it uses hardcoded predicates. This saves some time on parsing the math expression as well as drastically improves performance of range threshold.

<details><summary>Benchmark Results</summary>

Before the change:
```
goos: windows
goarch: amd64
pkg: github.com/grafana/grafana/pkg/expr
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
BenchmarkThreshold
BenchmarkThreshold/greater_than
BenchmarkThreshold/greater_than-16         	    3624	    319428 ns/op	  560673 B/op	   10061 allocs/op
BenchmarkThreshold/less_than
BenchmarkThreshold/less_than-16            	    3787	    313414 ns/op	  561320 B/op	   10060 allocs/op
BenchmarkThreshold/within_range
BenchmarkThreshold/within_range-16         	      37	  31990481 ns/op	 1674032 B/op	   30134 allocs/op
BenchmarkThreshold/within_range,_no_labels
BenchmarkThreshold/within_range,_no_labels-16         	       3	 384440567 ns/op	610835274 B/op	10020193 allocs/op
BenchmarkThreshold/outside_range
BenchmarkThreshold/outside_range-16                   	      37	  32272343 ns/op	 1679473 B/op	   30135 allocs/op
```
After:

```
goos: windows
goarch: amd64
pkg: github.com/grafana/grafana/pkg/expr
cpu: 11th Gen Intel(R) Core(TM) i9-11900H @ 2.50GHz
BenchmarkThreshold
BenchmarkThreshold/greater_than
BenchmarkThreshold/greater_than-16         	    4327	    254582 ns/op	  448385 B/op	    9001 allocs/op
BenchmarkThreshold/less_than
BenchmarkThreshold/less_than-16            	    4837	    248919 ns/op	  448385 B/op	    9001 allocs/op
BenchmarkThreshold/within_range
BenchmarkThreshold/within_range-16         	    4528	    269994 ns/op	  448385 B/op	    9001 allocs/op
BenchmarkThreshold/within_range,_no_labels
BenchmarkThreshold/within_range,_no_labels-16         	    4808	    292808 ns/op	  448385 B/op	    9001 allocs/op
BenchmarkThreshold/outside_range
BenchmarkThreshold/outside_range-16                   	    3859	    293570 ns/op	  448385 B/op	    9001 allocs/op
```

</details>

**Why do we need this feature?**
Improve the performance of the alerting evaluation. 

**Who is this feature for?**
Users with rules that query data sources that do not provide labels and apply range thresholds to the result.

**Special notes for your reviewer:**

Please check that:
- [x] It works as expected from a user's perspective.
- [x] If this is a pre-GA feature, it is behind a feature toggle.
- [x] The docs are updated, and if this is a [notable improvement](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/#how-to-determine-if-content-belongs-in-a-whats-new-document), it's added to our [What's New](https://grafana.com/docs/writers-toolkit/writing-guide/contribute-release-notes/) doc.